+++ src/DynamicFormsProvider.php
@@ -22,6 +22,10 @@ class DynamicFormsProvider extends ServiceProvider
             return new ComponentRegistry;
         });
 
+        $this->app->singleton(ResourceRegistry::class, function ($app) {
+            return new ResourceRegistry();
+        });
+
         $this->app->singleton(JSONLogicInitHelper::class, function ($app) {
             return new JSONLogicInitHelper;
         });
@@ -66,14 +70,17 @@ class DynamicFormsProvider extends ServiceProvider
 
         /** @var ComponentRegistry $registry */
         $registry = $this->app->make(ComponentRegistry::class);
-
         /** @var JSONLogicInitHelper $jsonHelper */
         $jsonHelper = $this->app->make(JSONLogicInitHelper::class);
         /** @var FileComponentRegistry $fileRegistry */
         $fileRegistry = $this->app->make(FileComponentRegistry::class);
 
-        Request::macro('validateDynamicForm', function (string $definitionJson, string $submissionJson) {
-            $formDefinition = new Form($definitionJson); // @TODO pass registry
+        Request::macro('validateDynamicForm', function (string $definitionJson, string $submissionJson, ?ResourceRegistry $resourceRegistry) {
+            if (is_null($resourceRegistry)) {
+                $resourceRegistry = resolve(ResourceRegistry::class);
+            }
+
+            $formDefinition = new Form($definitionJson, $resourceRegistry); // @TODO pass registry
             $validator = $formDefinition->validate($submissionJson);
 
             if (! $validator->isValid()) {