+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
+### Added
+- Support for Resources within the Select component.
+
 ## [v0.8.4 - 2022-07-26]
 ### Fixes
 - Allows for regular textarea editors in addition to Quill.